Fix bug with page navigation
When navigating to the next page, the number of shows in the next page
is calculated *after* the current index is increased.

Say you have a list of 7 shows, the number of shows at any given page
is 3, and as a result, you get three pages. When you navigate to the
last page, the max (current index) will equal 6, and the first visible
element index will be the first visible element; In this case, 4, since
the code renders from first visible index until number of shows.

When you navigate back to the second page, the first visible element is
going to equal 1, and the current index 3. When you scroll *back* to the
last page, the current index will equal 6, and so when slicing the
length will equal 1, and it will ONLY slide by one and increase the
first visible index by 1.

This fix implements a while loop to increase the first visible element's
index in any given category to be number of shows at any given page LESS
than the current index (in the unique case above), ensuring it slides by
the appropriate amount and doesn't mess up future calculations.