Fix bug when sliding back from 2nd page to 1st
Whenever you resize the screen, the current page's value is updated
to match the new page value based on number of results shown per page.
This causes a bug to happen in a specific case; when on the 2nd page,
and you resize to bigger, the current page's value is set to 1st page,
so when you scroll back (from 2nd page to 1st) it slides you back to the
last page (because it assumes you're on first, and want to go to last).