Add IDs and observer for hovering div
Vanilla JavaScript is a bit finicky when it comes to detecting mouse
events in the context of fast scrolling or rapid mouse movement.
The hovering show's div (the div that shows when you hover over a show)
mouseleave event sometimes did not properly fire when rapidly scrolling.
So, for each hovering div, a new observer is instantiated to check if
the hovering div went off-screen. If it did, remove it.