Fix hover div data retrieval and rendering
Separated the dom logic entirely from the application logic for the
hovering div. Now, when you hover, it shows you a loading skeleton,
and as soon as the controller retrieves the data from the model, it
calls an updateMetadata method, which inserts the data dynamically
into the DOM without the controller even knowing what div needs it.

Previously, the controller itself would activate hover even if the
mouse is no longer hovering over the show. If the mouse was still
hovering, it would still be an issue since (depending on the user's
internet) it would only render the hover element once data is retrieved.

Now, it gives the user an indicator that it's retrieving the elements.