Pause & resume billboard video when modal opens
This is important in one case only: When the user doesn't scroll down
enough for the billboard to pause before opening up a show's info.
Usually, this is only applicable when the user views one of the first
two categories' shows (in full screen size), but the code works for any
screen size.