Fix hovering bug on scroll
We're here again. This is a temporary fix because, again, JS is finnicky
when it comes to mouse events detection. When scrolling, sometimes, the
hovering div that is created to show the title's metadata remains up.
This new update attaches a scroll event to the document, so that when
the user scrolls, it disables the hovering event, and when the user's
mouse moves, it enables it. Remaining is to fix one more bug: When
enabling, if the mouse was already on a show you have to exit it and re-
enter.