Fix show not activating hover after scroll
The hover is disabled as the user scrolls so it doesn't incorrectly
register the mouse enter event only because the mouse came over the
element by chance as the user scrolled.

This lead to a bug that happens when the user stops scrolling and the mouse
is over the element they want, so if they move the mouse slightly hover
doesn't work because mouseenter needs the mouse to leave the element and
re-enter it.

This commit fixes that bug by checking once if the mouse is over a
show and activates the hover if it is.